Feature : Local DB Migration for Africa Table in SL         - Offline login issue has been fixed 	- My patient navigation handling according to role 	- MDT-16579, MDT-16630, MDT-16665

+++ Spice-SL/app/src/main/java/com/medtroniclabs/spice/ui/phuwalkins/viewmodel/PhuWalkInsViewModel.kt
@@ -73,8 +73,8 @@ class PhuWalkInsViewModel @Inject constructor(
             SecuredPreference.getString(DefinedParams.houseHoldLinkStartTiming)?.let {
                 householdMemberRepository.addLinkMemberCall(
                     memberID.toString(),
-                    callStartTime = it,
-                    callEndTime =  com.medtroniclabs.spice.app.analytics.utils.CommonUtils.getCurrentDateTimeInLocalTime()
+                    callStartTime = it.toLong(),
+                    callEndTime =  com.medtroniclabs.spice.app.analytics.utils.CommonUtils.getCurrentDateTimeInLocalTime().toLong()
                 )
             }
         }
